refactor for tsx
